// web/src/repositories/clients/private/models/AddGameSteps201Response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * tzetypes-badminton
 * tzetypes-badminton
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { GameStep } from './GameStep';
import {
    GameStepFromJSON,
    GameStepFromJSONTyped,
    GameStepToJSON,
} from './GameStep';

/**
 * 
 * @export
 * @interface AddGameSteps201Response
 */
export interface AddGameSteps201Response {
    /**
     * 
     * @type {Array<GameStep>}
     * @memberof AddGameSteps201Response
     */
    gameSteps: Array<GameStep>;
}

/**
 * Check if a given object implements the AddGameSteps201Response interface.
 */
export function instanceOfAddGameSteps201Response(value: object): value is AddGameSteps201Response {
    if (!('gameSteps' in value) || value['gameSteps'] === undefined) return false;
    return true;
}

export function AddGameSteps201ResponseFromJSON(json: any): AddGameSteps201Response {
    return AddGameSteps201ResponseFromJSONTyped(json, false);
}

export function AddGameSteps201ResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): AddGameSteps201Response {
    if (json == null) {
        return json;
    }
    return {
        
        'gameSteps': ((json['game_steps'] as Array<any>).map(GameStepFromJSON)),
    };
}

export function AddGameSteps201ResponseToJSON(value?: AddGameSteps201Response | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'game_steps': ((value['gameSteps'] as Array<any>).map(GameStepToJSON)),
    };
}