// web/src/repositories/clients/private/models/StartGame201Response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * tzetypes-badminton
 * tzetypes-badminton
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Game } from './Game';
import {
    GameFromJSON,
    GameFromJSONTyped,
    GameToJSON,
} from './Game';
import type { GameStep } from './GameStep';
import {
    GameStepFromJSON,
    GameStepFromJSONTyped,
    GameStepToJSON,
} from './GameStep';

/**
 * 
 * @export
 * @interface StartGame201Response
 */
export interface StartGame201Response {
    /**
     * 
     * @type {Array<GameStep>}
     * @memberof StartGame201Response
     */
    steps: Array<GameStep>;
    /**
     * 
     * @type {Game}
     * @memberof StartGame201Response
     */
    game: Game;
}

/**
 * Check if a given object implements the StartGame201Response interface.
 */
export function instanceOfStartGame201Response(value: object): value is StartGame201Response {
    if (!('steps' in value) || value['steps'] === undefined) return false;
    if (!('game' in value) || value['game'] === undefined) return false;
    return true;
}

export function StartGame201ResponseFromJSON(json: any): StartGame201Response {
    return StartGame201ResponseFromJSONTyped(json, false);
}

export function StartGame201ResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): StartGame201Response {
    if (json == null) {
        return json;
    }
    return {
        
        'steps': ((json['steps'] as Array<any>).map(GameStepFromJSON)),
        'game': GameFromJSON(json['game']),
    };
}

export function StartGame201ResponseToJSON(value?: StartGame201Response | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'steps': ((value['steps'] as Array<any>).map(GameStepToJSON)),
        'game': GameToJSON(value['game']),
    };
}